+++ src/utils.c
@@ -624,8 +624,7 @@ SEXP rnng_device(SEXP s1, SEXP s2) {
                       *(nng_socket *) R_ExternalPtrAddr(s2));
   if (xc)
     return mk_error(xc);
-
-  return Rf_ScalarInteger(0);
+  return Rf_ScalarInteger(xc);
 
 }