+++ NEWS.md
@@ -2,7 +2,7 @@
 
 #### New Features
 
-* Added `nng_clock()` and `nng_random()` utilities exposing the library functions for timing and cryptographic RNG respectively.
+* Added `nng_clock()`, `nng_sleep()` and `nng_random()` utilities exposing the library functions for timing and cryptographic RNG respectively.
 
 # nanonext 0.5.1